requirements.txt
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
click==8.1.7
colorama==0.4.6
fastapi==0.105.0
greenlet==3.0.2
h11==0.14.0
idna==3.6
mysql==0.0.3
mysql-connector-python==8.2.0
mysqlclient==2.2.1
protobuf==4.21.12
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
redis==4.6.0
sniffio==1.3.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
uvicorn==0.24.0.post1
